Design: Add <code> syntax highlighting
This is achieved through the highlight.js library, which can be
included in any template by doing the following:

{% block head_style_extras %}
{% stylesheet 'vendor_highlight' %}
{% endblock %}

which will load the monokai-sublime theme, and

{% block scripts %}
  {% javascript 'vendor_highlight' %}
  <script>
    hljs.highlightAll();
  </script>
{% endblock scripts %}

which will load the library and look for every <pre><code> element in
a page and perform static highlight.

If the library is not loaded, a pure monochrome css fallback is
available.